requirements.txt
aenum==2.0.8
amqp==1.4.9
anyjson==0.3.3
apispec==0.31.0
attrs==17.4.0
Automat==0.6.0
bcrypt==3.1.4
bunq-sdk==0.12.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
constantly==15.1.0
crochet==1.9.0
decorator==4.2.1
enum-compat==0.0.2
eventlet==0.21.0
greenlet==0.4.13
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
kombu==3.0.37
marshmallow==2.15.0
marshmallow-enum==1.4.1
mock==2.0.0
nameko==2.8.4
path.py==11.0
pbr==3.1.1
pluggy==0.6.0
psycopg2-binary==2.7.4
py==1.5.2
pycparser==2.18
pycryptodomex==3.4.6
PyJWT==1.5.3
pytest==3.4.1
PyYAML==3.12
requests==2.18.1
simplejson==3.11.1
six==1.11.0
SQLAlchemy==1.2.4
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
Twisted==17.9.0
urllib3==1.21.1
virtualenv==15.1.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.4.3
